eclipse-plugin/plugins/com.google.eclipse.javascript.jstestdriver.ui/src/com/google/eclipse/javascript/jstestdriver/ui/view/actions/FailureOnlyViewerFilter.java
/*
 * Copyright 2009 Google Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not
 * use this file except in compliance with the License. You may obtain a copy of
 * the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations under
 * the License.
 */
package com.google.eclipse.javascript.jstestdriver.ui.view.actions;

import org.eclipse.jface.viewers.Viewer;
import org.eclipse.jface.viewers.ViewerFilter;

import com.google.eclipse.javascript.jstestdriver.core.model.ResultModel;

/**
 * Filter for the Test Result tree view, which only shows failures.
 * 
 * @author shyamseshadri@gmail.com (Shyam Seshadri)
 */
public class FailureOnlyViewerFilter extends ViewerFilter {

  @Override
  public boolean select(Viewer viewer, Object parentElement, Object element) {
    ResultModel model = (ResultModel) element;
    return !model.passed();
  }
}
